+++ packages/playground/index.html
@@ -11,7 +11,7 @@
     <script type="module" src="/src/main.tsx"></script>
     <script>
       const options = {
-        mountedPath: "/api/fp",
+        mountedPath: "/fp",
         openapi: {
           // This config assumes you have a local server running on :8787 with the OpenAPI spec exposed at /openapi.json
           //
